+++ changes.md
@@ -2,10 +2,16 @@
 
 ## Beta 0.2.1
 This patch fixes a number of issues related to UI.
+* Cards
+    * Location now properly announced (currently for shop)
 * Potions
-    * Potion location is now properly announced (currently for main screen and combat rewards)
+    * Potion location is now properly announced (currently for main screen and, shop, combat rewards)
     * Fixed a bug where you could not read the tooltip for a nempty potion slot
-    
+* Relics
+    * Location is now properly announced (now for shops)
+* Shops
+    * The button to sell a card is now actually announced as a button and has location information
+
 ## Beta 0.2.0
 
 * The versioning scheme has been changed to semantic versioning. This should allow for more clear version numbering for frequent updates instead of large spread out beta releases as it has been so far.